fix(kernel): anchor stopped generation conversion
